workshop/intake/agent-bundles/crewai-firecrawl-scraper/requirements.txt
crewai>=0.28.8
crewai[tools]>=0.28.8
firecrawl-py>=0.0.16
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
langchain>=0.1.0
langchain-openai>=0.0.6
openai>=1.12.0